pyproject.toml
[project]
name = "python-agentframework-demos"
version = "0.1.0"
description = "Collection of Python examples for Microsoft Agent Framework"
requires-python = ">=3.10"
dependencies = [
    "azure-identity",
    "openai>=1.109.1",
    "python-dotenv",
    "pydantic",
    "rich",
    "dotenv-azd",
    "aiohttp",
    "faker",
    "fastmcp",
    "opentelemetry-exporter-otlp-proto-grpc",
    "azure-monitor-opentelemetry",
    "psycopg[binary]",
    "pgvector",
    "azure-ai-evaluation[redteam]>=1.15.0",
    "agent-framework-core @ git+https://github.com/microsoft/agent-framework.git@11628c3166a1845683c5aef1e0d389eb862bcbaa#subdirectory=python/packages/core",
    "agent-framework-devui @ git+https://github.com/microsoft/agent-framework.git@11628c3166a1845683c5aef1e0d389eb862bcbaa#subdirectory=python/packages/devui",
    "agent-framework-redis @ git+https://github.com/microsoft/agent-framework.git@11628c3166a1845683c5aef1e0d389eb862bcbaa#subdirectory=python/packages/redis",
    "agent-framework-mem0 @ git+https://github.com/microsoft/agent-framework.git@11628c3166a1845683c5aef1e0d389eb862bcbaa#subdirectory=python/packages/mem0",
    "agent-framework-azure-ai-search @ git+https://github.com/microsoft/agent-framework.git@11628c3166a1845683c5aef1e0d389eb862bcbaa#subdirectory=python/packages/azure-ai-search",
    "agent-framework-orchestrations @ git+https://github.com/microsoft/agent-framework.git@11628c3166a1845683c5aef1e0d389eb862bcbaa#subdirectory=python/packages/orchestrations",
]

[tool.uv]
prerelease = "allow"

[tool.ruff]
line-length = 120
target-version = "py310"
lint.select = ["E", "F", "I", "UP"]
lint.ignore = ["D203"]